Modifications and Implementations.
Intelledger (grails-app/domain):

* Added some additional getters, some of which are currently not used.

IntelledgerEntry:

* dos2unix (^M line-endings)
* Added another getter for readability.
* Altered the constraints so that `changeId` and `version` are **required**
fields, as to have that be optional defeats the purpose of Intelledger
(auditing).

Intelledger (src/groovy):

* Changed functions to move to a more "active-is-assumed" perspective;
that is, it is assumed in get, set, delete, etc. that these operations
are working with only the _active_ entries. No or few methods to
interact with all entries are present at this time.

* Method signatures reflect the required `changeId` field.

* Altered the `_set()` logic a little to shrink method size.

* Removed the ability to `delete()` an **Intelledger** itself, as there
is no way of auditing such an action and therefore defeats the purpose
of **Intelledger** to do so.

* Altered the behavior of the `delete()` functions, such that rather
than actually deleting the last entry, a new one is created with the
required `changeId` of the person who deleted it, with the same
information as the previous entry; both are marked _inactive_,
effectively rendering them deleted from all usual methods.

* JavaDoc updates (needs more updating, but this is at least a start).